fix(annotation-tests): conditionally import logging context reliant annotation helpers
